requirements.txt
uvicorn
fastapi
Pillow
pickle
torch
